Minor improvements

+++ rpi/gf.sh
@@ -19,17 +19,21 @@ cd /home/pi
          sudo rm -f background.log
          echo yes   >background.log
          sudo chmod 777 background.log
-
   fi
+  
 
-sudo rm -f gf.tmp
-sudo cp -f gf.log gf.tmp
-sudo rm -f gf.log
+ if [ -e "/gf.log" ]
+  then
+        sudo rm -f gf.tmp
+        sudo cp -f gf.log gf.tmp
+        sudo rm -f gf.log
+ fi
 
 echo submiting gForth  # Using one of the following lines:
 # sudo nice --10 nohup gforth-fast _demo1.f        1>gf.log 2>&1 &
 # sudo nice --10 nohup gforth-fast _UploadServer.f 1>gf.log 2>&1 &
-sudo nohup gforth-fast _SensorWeb1.fs  1>gf.log 2>&1  & 
+# sudo nice --10 nohup gforth-fast _DemoMaster.fs  1>gf.log 2>&1 &
+  sudo nohup gforth-fast _SensorWeb1.fs            1>gf.log 2>&1  &
 
 exit 0
